% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalBoundingBox.R
\name{boundingBox}
\alias{boundingBox}
\title{Rectangle area around points}
\usage{
boundingBox(points, buffer = 0, epsg = 4326)
}
\arguments{
\item{points}{a sf object, points}

\item{buffer}{buffer distance between the points and the rectangle; defaults 0}

\item{epsg}{projection of the output polygon}
}
\value{
sf polygon
}
\description{
Creates optimal rectangle bounding box around points
}
\details{
The code is based on a Rotating Caliper Algorithm and mostly copy and pasted (see reference)
}
\references{
http://dwoll.de/rexrepos/posts/diagBounding.html
}
\author{
Marvin Ludwig
}
