% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropSentinel.R
\name{cropSentinel}
\alias{cropSentinel}
\title{Crop Sentinel}
\usage{
cropSentinel(pathL2A, cropExtent, outpath)
}
\arguments{
\item{pathL2A}{filepath to L2A Sentinel data}

\item{cropExtent}{sf / sp crop extent}

\item{outpath}{filepath for the cropped scene}
}
\value{

}
\description{
Crops Sentinal data to mof extent
}
\author{
Marvin Ludwig
}
