% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown.radius.R
\name{crown.radius}
\alias{crown.radius}
\title{Calculate Tree Crown Radius}
\usage{
crown.radius(crowns, stems = NULL)
}
\arguments{
\item{crowns, }{tree crown geometries}

\item{stems, }{stem positions}
}
\value{
vector with radi
}
\description{
Distance from stem position to farthest crown polygon vertex
}
\details{
If no stems are provided, the polygon centroids are used as stem positions
}
\author{
Marvin Ludwig
}
