% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsFunctions.R
\name{gpsPretty}
\alias{gpsPretty}
\title{Rename GPS export colnames}
\usage{
gpsPretty(gpsExport)
}
\arguments{
\item{gpsExport}{data.frame, the XPAD export using the GNSSESSENTIAL template}
}
\value{
the gpsExport data.frame with pretty column names
}
\description{
Rename GPS export colnames
}
\details{
Renames a GNSSESSENTIAL export from XPAD

Makes you less likely to jump out of the window because of dots and weird abbreviations in colnames.
}
