% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.split.R
\name{raster.split}
\alias{raster.split}
\title{Raster Class Splitter}
\usage{
raster.split(x)
}
\arguments{
\item{x}{raster}
}
\value{
raster stack with one layer per class
}
\description{
Takes a raster with categorial values and splits it into binary layers
}
\author{
Marvin Ludwig
}
