% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSentinel.R
\name{readSentinel}
\alias{readSentinel}
\title{Read Sentinel}
\usage{
readSentinel(pathL2A)
}
\arguments{
\item{pathL2A}{filepath to L2A Sentinel data}
}
\value{
raster stack
}
\description{
reads Sentinel Stack from .SAFE
}
\author{
Marvin Ludwig
}
