% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel.download.R
\name{sentinel.download}
\alias{sentinel.download}
\title{Download L2A Sentinel data}
\usage{
sentinel.download(username, password, area, date1, date2, outpath)
}
\arguments{
\item{username}{Copernicus username}

\item{password}{Copernicus password}

\item{area}{study area as a path geojson}

\item{date1}{start search date YYYYMMDD}

\item{date2}{end search date YYYYMMDD}

\item{outpath}{directory for the imagery}
}
\value{

}
\description{
Query and download Sentinel-2 data from the Copernicus hub
}
\author{
Marvin Ludwig
}
