% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentinel.read.R
\name{sentinel.read}
\alias{sentinel.read}
\title{Read Sentinel}
\usage{
sentinel.read(pathL2A)
}
\arguments{
\item{pathL2A}{filepath to L2A Sentinel data}
}
\value{
raster stack
}
\description{
reads Sentinel Stack from .SAFE
}
\author{
Marvin Ludwig
}
