% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.pattern.R
\name{sf.pattern}
\alias{sf.pattern}
\title{Simple feature to spatstat pointpattern}
\usage{
sf.pattern(p, window, mark = NULL)
}
\arguments{
\item{p}{sf points}

\item{window}{sf polygon: the observation window (e.g. study area)}

\item{mark}{character vector, marks for each point (optional)}
}
\value{
a spatial point pattern
}
\description{
Converts simple feature points to a spatial point pattern
}
\author{
Marvin Ludwig
}
