% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crown.attributes.R
\name{tree.attributes}
\alias{tree.attributes}
\title{Calculate tree hulls with crown attributes}
\usage{
tree.attributes(tree_pc, stems = NULL)
}
\arguments{
\item{tree_pc, }{classified tree}

\item{stems, }{stem positions for crown radius (optional)}
}
\value{

}
\description{
height related attributes
}
\author{
Marvin Ludwig
}
