% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crownShape.R
\name{crownShape}
\alias{crownShape}
\title{Crown Shapes}
\usage{
crownShape(tree, distance, angle, crs = 25832)
}
\arguments{
\item{tree}{sf point tree}

\item{distance}{vector with measured distances}

\item{angle}{vector with measured angles}

\item{crs}{epsg code}
}
\value{
tree with crown shape poylgon
}
\description{
Adds crown geometry to tree
}
\author{
Marvin Ludwig
}
